capnp: quick sketch by @schwa to convert mavlink msgs to capping msgs, in order to investigate their size
NB: `common_annotated.capnp` is the output of `capnp compile -ocapnp common.capnp`, and the struct size annotations specify the size of the fields themselves, without taking into account any additional serialization info (per https://capnproto.org/encoding.html#serialization-over-a-stream)

we could potentially get away with smaller serialization data if we could live with non-word-aligned data in the serialization header (i think we could).

more radically, we could also reduce size even further if we mandated that all messages must fit in a single ‘segment’, such that it’s not necessary to specify the segment count, and the size of the single segment is simply the size of the received data (i.e., the size of the received datagram over UDP). this would reduce compatibility with other full-fledged capn proto implementations in the world, so may not be worth it, but is a feasible way of wringing out at least 8 more bytes (maybe more with padding) from each msg.